update stub killer

+++ L1Trigger/TrackFindingTracklet/interface/StubKiller.h
@@ -23,12 +23,12 @@ public:
   bool killStub(
   		const TTStub<Ref_Phase2TrackerDigi_>* stub,
 		const vector<int> layersToKill,
-		const int minPhiToKill,
-		const int maxPhiToKill,
-		const int minZToKill,
-		const int maxZToKill,
-		const int minRToKill,
-		const int maxRToKill,
+		const double minPhiToKill,
+		const double maxPhiToKill,
+		const double minZToKill,
+		const double maxZToKill,
+		const double minRToKill,
+		const double maxRToKill,
 		const double fractionOfStubsToKillInLayers,
 		const double fractionOfStubsToKillEverywhere
   	);
@@ -48,12 +48,12 @@ private:
 	const TrackerGeometry* trackerGeometry_;
 
 	vector<int> layersToKill_;
-	int minPhiToKill_;
-	int maxPhiToKill_;
-	int minZToKill_;
-	int maxZToKill_;
-	int minRToKill_;
-	int maxRToKill_;
+	double minPhiToKill_;
+	double maxPhiToKill_;
+	double minZToKill_;
+	double maxZToKill_;
+	double minRToKill_;
+	double maxRToKill_;
 	double fractionOfStubsToKillInLayers_;
 	double fractionOfStubsToKillEverywhere_;
 	double fractionOfModulesToKillEverywhere_;
